Authority Refactor - Stab One
