+++ lib/msal-core/src/UserAgentApplication.ts
@@ -5,9 +5,9 @@ import { AccessTokenCacheItem } from "./AccessTokenCacheItem";
 import { AccessTokenKey } from "./AccessTokenKey";
 import { AccessTokenValue } from "./AccessTokenValue";
 import { ServerRequestParameters } from "./ServerRequestParameters";
-import { Authority } from "./Authority";
+import { Authority } from "./authority/Authority";
 import { ClientInfo } from "./ClientInfo";
-import { Constants, SSOTypes, PromptState, BlacklistedEQParams, InteractionType, libraryVersion } from "./utils/Constants";
+import { Constants, InteractionType, libraryVersion } from "./utils/Constants";
 import { IdToken } from "./IdToken";
 import { Logger } from "./Logger";
 import { Storage } from "./Storage";
@@ -19,9 +19,9 @@ import { TokenUtils } from "./utils/TokenUtils";
 import { TimeUtils } from "./utils/TimeUtils";
 import { UrlUtils } from "./utils/UrlUtils";
 import { ResponseUtils } from "./utils/ResponseUtils";
-import { AuthorityFactory } from "./AuthorityFactory";
+import { AuthorityFactory } from "./authority/AuthorityFactory";
 import { Configuration, buildConfiguration, TelemetryOptions } from "./Configuration";
-import { AuthenticationParameters, validateClaimsRequest } from "./AuthenticationParameters";
+import { AuthenticationParameters } from "./AuthenticationParameters";
 import { ClientConfigurationError } from "./error/ClientConfigurationError";
 import { AuthError } from "./error/AuthError";
 import { ClientAuthError, ClientAuthErrorMessage } from "./error/ClientAuthError";
@@ -1551,7 +1551,7 @@ export class UserAgentApplication {
           scopes: [],
           expiresOn: null,
           account: null,
-          accountState: "",
+          accountState: ""
       };
 
       let error: AuthError;
@@ -1629,13 +1629,8 @@ export class UserAgentApplication {
                       response = ResponseUtils.setResponseIdToken(response, idTokenObj);
                   }
 
-                  // retrieve the authority from cache and replace with tenantID
-                  const authorityKey = Storage.generateAuthorityKey(stateInfo.state);
-                  let authority: string = this.cacheStorage.getItem(authorityKey, this.inCookie);
-
-                  if (!StringUtils.isEmpty(authority)) {
-                      authority = UrlUtils.replaceTenantPath(authority, response.tenantId);
-                  }
+                  // set authority in response
+                  const authority: string = this.setAuthority(stateInfo.state, this.inCookie, this.cacheStorage, idTokenObj);
 
                   // retrieve client_info - if it is not found, generate the uid and utid from idToken
                   if (hashParams.hasOwnProperty(Constants.clientInfo)) {
@@ -1695,12 +1690,8 @@ export class UserAgentApplication {
                       this.logger.warning("ClientInfo not received in the response from AAD");
                   }
 
-                  authorityKey = Storage.generateAuthorityKey(stateInfo.state);
-                  let authority: string = this.cacheStorage.getItem(authorityKey, this.inCookie);
-
-                  if (!StringUtils.isEmpty(authority)) {
-                      authority = UrlUtils.replaceTenantPath(authority, idTokenObj.tenantId);
-                  }
+                  // set authority in response
+                  const authority: string = this.setAuthority(stateInfo.state, this.inCookie, this.cacheStorage, idTokenObj);
 
                   this.account = Account.createAccount(idTokenObj, new ClientInfo(clientInfo));
                   response.account = this.account;
@@ -1760,6 +1751,28 @@ export class UserAgentApplication {
           throw AuthError.createUnexpectedError("Response is null");
       }
       return response;
+  }
+
+  /**
+     * Set Authority when saving Token from the hash
+     * @param state
+     * @param inCookie
+     * @param cacheStorage
+     * @param idTokenObj
+     * @param response
+     */
+  private setAuthority(state: string, inCookie: boolean, cacheStorage: Storage, idTokenObj: IdToken): string {
+      const authorityKey: string = Storage.generateAuthorityKey(state);
+      const cachedAuthority: string = cacheStorage.getItem(authorityKey, inCookie);
+
+      let specificAuthority: string = cachedAuthority;
+
+      // retrieve the authority from cache and replace with tenantID
+      if (!StringUtils.isEmpty(cachedAuthority)) {
+          specificAuthority = UrlUtils.replaceTenantPath(cachedAuthority, idTokenObj.tenantId);
+      }
+
+      return specificAuthority;
   }
   /* tslint:enable:no-string-literal */
