+++ lib/msal-core/src/authority/IInstanceDiscoveryResponse.ts
